server/retrain/requirements.txt
Pillow==6.0.0
pkg-resources==0.0.0
